barista: Extract shared interface from environment files.
To avoid problems during the angular environment replacements, we have
extracted the interface to its own file.

Fixes #569
