+++ tools/environments/barista-environment.interface.ts
@@ -1,0 +1,29 @@
+/**
+ * @license
+ * Copyright 2020 Dynatrace LLC
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+export interface BaEnvironment {
+  rootDir: string;
+  distDir: string;
+  examplesMetadataDir: string;
+  examplesMetadataFileName: string;
+  examplesLibDir: string;
+  demosAppDir: string;
+  baristaAppDir: string;
+  iconsRoot: string;
+  iconsChangelogFileName: string;
+  strapiEndpoint: string | undefined;
+  internalLinks: string | undefined;
+}
